Derived mapping spreadsheet for analysing reversible reactions
FakeReversible.csv
Made a CSV file containing the reactions that are `not actually
reversible', see last commit

StocReacs.csv
Made a CSV file derived from the latest mapping spreadsheet
containing only the name of the reactions as per the stoichiometric
model and the relationship/type of reaction.  This is so that I
can use this with the results from analysing the bounds from the
objective exercise for further analysis.  More specifically,
resolving which reactions are actually reversible.