Attempted to add 41-dimension boundaries as a table in appendix
Attempted to do so using PGF as in my Part II dissertation.
For some reason it thinks I have an incomplete \ifx even though
NOWHERE in my TEX file is there an \if, and it adds an \fi, and it
just makes its problem worse. Why the fuck does it do this