Re-structured introduction: background-aims-methods
This conforms to the usual order of content used in papers (and
my Part II disseration too). The thing is that the lengths of these
bits are VERY different from my dissertation. Specifically, the
background is much shorter, and the methods are a lot longer.
Potentially I can move some of the `methods' to the main part of
the report. However, I don't think it's *that* out of place
among scientific papers. I must say I like this structure more

Also there's the last paragraph about unit conversions, which can
easily be moved to the discussion...