extended report: adding diagram to explain ppp mapping and added more explanation
as per jorge's comments:

The mapping and the way in which it has been done is an important
contribution; thus, we should provide more details. Clarification
is provided:
   (1) Which metabolite yields each rule (e.g. X5P the first rule)
   (2) The 'equals sign' means equating flux bounds and sums of
       flux bounds

This clarification is provided by the diagram added, the brackets
after each mapping rule devised, and a paragraph to explain the
sum sign.