made comflux() able to compute flux of any reaction
purpose: find the equivalent of one-reaction plots for every
reaction in the kinetic model, but instead of citramalate
productivity i want fluxes through each reaction

This is so that I have separate lists for each reaction in the
kinetic model. Thus, computing flux bounds (8d, or otherwise) can
use different sets of reactions in differential evolution,
potentially making the bounds better.

I've also modified vmax_noloop.py so that incorporates comflux()
rather than the usual comproducti(), as I did to generate the
figure in the discussion to address the validity of using flux
to approximate productivity