addressing lund medium according to chassagnole et al 2005
Previously the method I used to use concentrations in the Lund
medium to specify bounds for exchange reactions in the
stoichiometric model (as written in the extended report) produces
very low values.  This might be because the OD600 used in the
equation does not translate well to the conditions tested.
I derived an equation based on the model equations in Chassagnole
et al. (2005):

Rmax = ρx.D . Cglc
       ----
        Cx

Where:
- Rmax = maximum glucose uptake rate
- ρx = specific weight of biomass
- D = dilution rate
- Cx = biomass concentration
- Cglc = concentration of glucose in feed medium