extended report: edits to finish for now
- general clarity/wording/spelling corrections, diagram positions,
  added general comments
- made captions smaller (stylistic choice)
- fixed labelling of some equations
- add mention of fork in abstract. it is a major point that needs
  further addressing
- commented out sentence about stoichiometric model structure: not
  sure if fact is true, in any case it's poor practice
- commented out section on inflections points -- not important and
  not mentioned later
- commented out section on glycolytic enzymes -- conclusion does
  not relate to anywhere else
- shortened conclusion of section on glucose uptake -- implications
  not used, is blagging anyway, and in any case better in discussion
- (lb, ub) -> (LB, UB) is a stylistic choice
- outflux -> efflux: word choice (outflux isn't a word)
- revived the future direction section and updated it -- serves as
  a note to myself about what to do next
- updated dates in which i worked on the project

The extended report (as it stands) serves as a bridge between the
old main report and a potential manuscript.  It contains new
material from investigations in summer 2019 and clarifies some
things from summer 2018.  Most of the `meat' is in the stoichio-
metric model section, while most of the kinetic model section
just contains preliminary `interesting' investigations that are
a bit far from the intended main points of a potential manuscript.
So in this way, the extended report serves as a collection of
ideas and findings and I can choose which to put in a potential
manuscript later.